web/app/requirements.txt
celery==5.3.4
Flask==3.0.0
Flask_Mail==0.9.1
flask_paginate==2023.10.24
Flask_Security_Too==5.3.2
flask_wtf==1.2.1
python-slugify==8.0.1
SQLAlchemy==2.0.21
SQLAlchemy_Utils==0.41.1
Werkzeug==3.0.1
WTForms==3.0.1
